requirements.txt
django
psycopg2-binary
djangorestframework
Pillow
telebot
requests